+++ tests/Resources/InvalidTestNotification.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Kreait\Laravel\Firebase\Tests\Resources;
+
+use Illuminate\Notifications\Notification;
+
+class InvalidTestNotification extends Notification
+{
+}